Update backup creation flow
Currently, during the backup creation, the BackupManager persists
the backup driver type on database before instantiating the backup
service using that driver. This flow works fine with upstream
cinder backup drivers, because for those drivers the configuration
is provided via cinder.conf file, and it's fully validated before
any backup operation, during the cinder-backup startup.
But, with multitenant nfs driver, this backup creation flow doesn't
work very well, because the nfs hostpath is provided by the user,
using the backup location parameter, and if the backup hostpath
is invalid, the driver instantiation will fail, causing the backup
creation to also fail, leaving the backup on an error state.
Trying to delete the same backup later will also fail, since
the backup hostpath is the same, and BackupManager will again fail
to instantiate the backup service.
So, in order to avoid this issue when creating and deleting backups,
we are changing the backup creation flow to first try to instantiate
the backup service, and if it succeeds, update the backup driver
type on cinder database.
With this change, if the backup service instantiation fails during
the backup creation, the driver type won't be persisted, and
BackupManager won't try to instantiate the backup service later,
when the same backup is being deleted.

Jira: CGTS-22924

Signed-off-by: Daniel Pereira <Daniel.Pereira@windriver.com>

[ Ported commit on top of Cinder @ 2025.1 Epoxy ]
Signed-off-by: Andre de Held Badur <andre.deheldbadur@windriver.com>
(cherry picked from commit b22ab3274a1c3debacbd43c3be4025f548d79294)

Change-Id: I7a2662691948528d6cbf1bbf7340f109e1b8abd5